updates to grid automating codes (without comments)

+++ amber/examples/submit_multiple_jobs.py
@@ -6,7 +6,7 @@ import shutil
 
 sim_dirs = glob('/pscratch/sd/d/doughty/emuinf/grid/nyx*')
 
-target_successes = 7
+target_successes = 3
 success_counter = 0
 
 for directory in sim_dirs:
@@ -19,11 +19,16 @@ for directory in sim_dirs:
     if not os.path.exists(f"{directory}/zhi_H"):
         continue
 
+    # check if the job to run nyx is already submitted
+    if os.path.exists(f"{directory}/RUN_JOB_SUBMITTED"):
+        continue
+
     pltlist = glob(f"{directory}/plt*")
     # second check if there are already plt files
     if len(pltlist) > 0:
         continue
     else:
+        subprocess.Popen(['touch', f'{directory}/RUN_JOB_SUBMITTED'])
         shutil.copyfile('/pscratch/sd/d/doughty/emuinf/templates/run_nyx.saul', f"{directory}/run_nyx.saul")
         os.chdir(f"{directory}")
         subprocess.Popen(['sbatch', 'run_nyx.saul'])
